+++ src/resdk/resources/sample.py
@@ -4,6 +4,7 @@ import logging
 from resdk.shortcuts.sample import SampleUtilsMixin
 
 from ..utils.decorators import assert_object_exists
+from .background_task import BackgroundTask
 from .collection import BaseCollection, Collection
 
 
@@ -245,7 +246,8 @@ class Sample(SampleUtilsMixin, BaseCollection):
             sample.
         :return: Duplicated sample
         """
-        duplicated = self.api().duplicate.post(
+        task_data = self.api().duplicate.post(
             {"ids": [self.id], "inherit_collection": inherit_collection}
         )
-        return self.__class__(resolwe=self.resolwe, **duplicated[0])
+        background_task = BackgroundTask(resolwe=self.resolwe, **task_data)
+        return self.resolwe.sample.get(id__in=background_task.result())